chore: refactor and unit tests fix
