+++ jest.stub.js
@@ -12,7 +12,7 @@ global.mockMap = {
     getMapGL: () => mockMapGL,
     setHoverState: jest.fn(),
     getBeforeLayerId: jest.fn(),
-    styleIsLoaded: jest.fn(),
+    styleIsLoaded: () => true,
 }
 
 // https://stackoverflow.com/questions/57943736/how-to-fix-window-url-createobjecturl-is-not-a-function-when-testing-mapbox-gl
