+++ src/utils/numbers.js
@@ -15,7 +15,3 @@ export const setPrecision = value =>
     numberPrecision(getPrecision(Math.abs(value)))(value)
 
 export const kmToMiles = value => value * 0.621371192
-
-export const squareMetersToHectares = value => value / 10000
-
-export const squareMetersToAcres = value => value / 4046.8564224
